% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createEset.R
\name{createEset}
\alias{createEset}
\title{Create ExpressionSet object}
\usage{
createEset(expr_mat, pheno)
}
\arguments{
\item{expr_mat}{an expression matrix of class matrix, or data frame that can be converted to matrix.}

\item{pheno}{a dataframe corresponding to the expression matrix. The row names of pheno must be identical to column names of expr_mat.}
}
\value{
An ExpressionSet object
}
\description{
This function creates ExpressionSet from an expression matrix and a phenotype data frame
}
\examples{
create_eset(exprs_dmap,pData_dmap)
}
