% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_dmap.R
\docType{data}
\name{eset_dmap}
\alias{eset_dmap}
\title{ExpressionSet object for scaffold}
\format{
an ExpressionSet object
}
\source{
Kindly offered by Fred
}
\usage{
eset_dmap
}
\description{
Dmap records expression matrix of normal hematopoietic cells.
}
\keyword{datasets}
