% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eset_ilaria.R
\docType{data}
\name{eset_ilaria}
\alias{eset_ilaria}
\title{ExpressionSet object as new samples}
\format{
an ExpressionSet object
}
\source{
Kindly offered by Fred
}
\usage{
eset_ilaria
}
\description{
Ilaria records expression matrix of erythroid leukemia.
}
\keyword{datasets}
