% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprs_dmap}
\alias{exprs_dmap}
\title{Expression matrix for scaffold}
\format{
a matrix
}
\source{
Kindly offered by Fred
}
\usage{
data("exprs_dmap")
}
\value{
a matrix
}
\description{
exprs_map records expression matrix of normal
hematopoietic cells. The data is measured by microarray.
}
\keyword{datasets}
