% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprs_ilaria}
\alias{exprs_ilaria}
\title{Expression matrix as new samples}
\format{
a data frame
}
\source{
Kindly offered by Fred
}
\usage{
exprs_ilaria
}
\description{
exprs_ilaria records expression matrix of Acute Myeloid Leukemia of erythroid subtype.
}
\keyword{datasets}
