% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_de_genes.R
\name{find_de_genes}
\alias{find_de_genes}
\title{Find Differentially Expressed Genes}
\usage{
find_de_genes(eset, group, pval_cutoff = 0.05, lfc_cutoff = 2)
}
\arguments{
\item{eset}{an ExpressionSet}

\item{group}{a column name of pData(eset)}

\item{pval_cutoff}{a cutoff value for p value when selecting differentially expressed genes. By default 0.05}

\item{lfc_cutoff}{a cutoff value for logFC when selecting differentially expressed genes. By default 2}
}
\value{
A vector of names of differentially expressed genes
}
\description{
This function performs DE analysis using limma to find DE genes
}
\examples{
find_de_genes(eset_dmap,"cell_types")
}
