% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_name_mapper_hs}
\alias{gene_name_mapper_hs}
\title{Gene name mapper for human}
\format{
a data frame
}
\source{
downloaded from biomaRt
}
\usage{
gene_name_mapper_hs
}
\description{
This table contains ensembl_gene_id of all human gene, and corresponding ensembl_transcript_id, entrezgene_id ,hgnc_symbol, and refseq_mrna
}
\keyword{datasets}
