% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPlot.R
\name{loadingPlot}
\alias{loadingPlot}
\title{Generate a loading plot of scaffold PCA}
\usage{
loadingPlot(
    space,
    num_genes = 3,
    gene_name = "hgnc_symbol",
    angle = 30,
    df_only = FALSE
    )
}
\arguments{
\item{space}{A \code{scaffoldSpace} object created by function
\code{\link{buildScaffold}}.}

\item{num_genes}{Optional parameter indicating number of genes to be shown.}

\item{gene_name}{Character indicating which type of gene name to show by
the side of arrows. By default HGNC symbol is shown.}

\item{angle}{A number for the degree of rotation of labels on the plot.}

\item{df_only}{The data.frame for the loading plot will be returned
instead of the plot.}
}
\value{
A data frame indicating the loading scores of the genes
that contribute most to the selected principle components.
The loading plot is printed automatically, thus not returned.
}
\description{
This function takes a \code{\link{scaffoldSpace-class}} object
and generates a loading plot to visualize the genes
that contribute most to the selected principle components.
}
\examples{
utils::data("exprs_dmap", "pData_dmap", package = "spaceRAT")
space <- buildScaffold(exprs_dmap, pData_dmap, "cell_types")
loadingPlot(space)
}
