% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pData_dmap.R
\docType{data}
\name{pData_dmap}
\alias{pData_dmap}
\title{Phenotype data for scaffold}
\format{
a data frame
}
\source{
Kindly offered by Fred
}
\usage{
pData_dmap
}
\description{
pData_map records phenotype information corresponding to exprs_dmap.
}
\keyword{datasets}
