% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pData_ilaria}
\alias{pData_ilaria}
\title{Phenotype data for new samples}
\format{
a data frame
}
\source{
Kindly offered by Fred
}
\usage{
data(pData_ilaria)
}
\description{
pData_ilaria contains phenotype data corresponding to exprs_ilaria
}
\keyword{datasets}
