% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScaffold.R
\name{plotScaffold}
\alias{plotScaffold}
\title{Plot a scaffoldSpace object}
\usage{
plotScaffold(space, title = "Scaffold plot")
}
\arguments{
\item{space}{A scaffoldSpace object returned by buildScaffold function.}

\item{title}{Title for the plot}
}
\value{
ggplot object.
}
\description{
This function plot a scaffoldSpace object created by \code{buildScaffold}
function.
}
\examples{
scaffold <- buildScaffold("prebuilt_DMAP")
plotScaffold(scaffold, "Scaffold plot title")
}
