% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictSample.R
\name{predictSample}
\alias{predictSample}
\title{Measures the distances to the centroids of existing scaffold labels to new sample(s)}
\usage{
predictSample(
  projection,
  centroid_calculation = "mean",
  dims = c(1, 2),
  temperature = 1e+05,
  display_numbers = FALSE,
  exclude_small_prob = FALSE,
  min_prob_sum = NULL,
  title_centroid_projection = "Centroids projected onto scaffold PCA",
  title_distance_pheatmap = "Distance of Sample to Scaffold Centroids",
  title_probability_pheatmap =
    "Softmax of Rank Distances of Sample to Scaffold Centroids",
  verbose = TRUE,
  annotation = "ensembl_gene"
)
}
\arguments{
\item{projection}{A scaffold space object, returned by function
\code{\link{projectSample}}. Recommended to have run with \code{subset_interaction = TRUE}}

\item{centroid_calculation}{Calculate centroids by \code{'mean'} or \code{'median'} of class expressions.}

\item{dims}{A numeric vector containing 2 numbers, indicating
which two principle components to plot.}

\item{temperature}{Parameter for softmax function used transform distance to probability. 
Larger temperature → smoother probabilities.
Smaller temperature → peakier, closer to winner-takes-all.}

\item{display_numbers}{Display numbers in tiles of pheatmap (see \code{\link{pheatmap::pheatmap}}).}

\item{exclude_small_prob}{Logical. If TRUE, labels with low overall 
probabilities will be excluded from the results. Requires 
\code{min_prob_sum} to be specified.}

\item{min_prob_sum}{Numeric in the range [0, 1]. Threshold for excluding 
labels when \code{exclude_small_prob = TRUE}. Labels with a 
total summed probability across all samples less than this value will 
be removed. Defaults to \code{NULL}, and must be provided if 
\code{exclude_small_prob = TRUE}.}

\item{verbose}{A logical vector indicating whether to report the number of
genes imputed to make \code{sample} compatible with
\code{scaffold}}

\item{annotation}{Type of gene identifier to use for scaffold.
Currently "ensembl_gene", "ensembl_transcript", "entrez", "hgnc_symbol",
and "refseq_mrna" are supported.}

\item{title}{Title of the plot.}
}
\value{
ggplot object with new samples projected to existing scaffold plot
}
\description{
This function takes in a scaffold space objects, subsets
the new dataset to contain only the genes that defines the scaffold space,
ranks the subset within new sample(s),
finally projects the new sample(s) onto the existing scaffold PCA plot.
}
\examples{
utils::data("ilaria_counts", "ilaria_pData", package = "spaceRATScaffolds")
scaffold <- buildScaffold("DMAP")
# or create your own scaffoldSpace
p <- projectSample(scaffold,ilaria_counts,ilaria_pData,"cancer_type")
}
