% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSample.R
\name{projectSample}
\alias{projectSample}
\title{Project new sample(s) onto the existing scaffold PCA plot}
\usage{
projectSample(
  list_from_buildScaffold,
  eset_sample,
  group_sample,
  title = "Samples projected on scaffold PCA",
  verbose = T
)
}
\arguments{
\item{list_from_buildScaffold}{the list returned by function \code{buildScaffold()}}

\item{eset_sample}{the sample(s) to project, of class ExpressionSet}

\item{group_sample}{a column name of pData(eset_sample)}

\item{title}{Title of the plot}

\item{verbose}{a logical vector indicating whether to report the number of genes added to eset_sample to make it compatible with eset_scaffold}
}
\value{
a ggplot object with new samples projected to existing scaffold PCA plot
}
\description{
This function takes in objects returned from function \code{buildScaffold()}, subsets the new dataset, ranks the subset, finally projects the new sample(s) onto the existing scaffold PCA plot.
}
\examples{
projectSample(l,eset_sample,"cancer_type")
}
