% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_scaffoldSpace.R
\docType{class}
\name{scaffoldSpace-class}
\alias{scaffoldSpace-class}
\alias{scaffoldSpace}
\title{A S4 class to store all the information of scaffold PCA space}
\description{
A S4 class to store all the information of scaffold PCA space
}
\section{Slots}{

\describe{
\item{\code{DEgenes}}{a vector containing names of differentially expressed genes in scaffold dataset}

\item{\code{graph}}{a gg object returned by ggplot}

\item{\code{pca}}{a prcomp object returned by function \code{prcomp()}}
}}

