% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScaffold.R
\name{buildScaffold}
\alias{buildScaffold}
\title{Returns a ggplot and PCA parameters for ExpressionSet object}
\usage{
buildScaffold(
  eset_scaffold,
  group_scaffold,
  pval_cutoff = 0.05,
  lfc_cutoff = 2,
  title = "Scaffold PCA Plot",
  pca_scale = FALSE
)
}
\arguments{
\item{eset_scaffold}{an ExpressionSet}

\item{group_scaffold}{a column name of pData(eset_scaffold)}

\item{pval_cutoff}{a cutoff value for p value when selecting differentially expressed genes. By default 0.05}

\item{lfc_cutoff}{a cutoff value for logFC when selecting differentially expressed genes. By default 2}

\item{pca_scale}{a logical variable determining whether to normalize rows when plotting PCA}
}
\value{
A list containing a ggplot object and a prcomp object
}
\description{
This function performs differential expression analysis to select DE gene, then subset the scaffold dataset, rank the subset, finally produce a PCA plot using ranks.
}
\examples{
buildScaffold(eset_dmap,"cell_types")
buildScaffold(eset_dmap,"cell_types", pval_cutoff=0.01,pca_scale=T)
}
