% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildScaffold.R
\name{buildScaffold}
\alias{buildScaffold}
\title{Returns either a prebuilt scaffoldSpace object,
or a new one calculated from input expression matrix}
\usage{
buildScaffold(
    object,
    pheno = NULL,
    colname = NULL,
    assay = NULL,
    data = NULL,
    threshold = 10,
    add_umap = FALSE,
    classes = NULL,
    pval_cutoff = 0.05,
    lfc_cutoff = 2,
    pca_scale = FALSE,
    annotation = "ensembl_gene"
    )
}
\arguments{
\item{object}{An count or expression matrix of class
\code{\link[SummarizedExperiment:SummarizedExperiment]{SummarizedExperiment}}
,
\code{\link[SingleCellExperiment:SingleCellExperiment]{SingleCellExperiment}}
, \code{matrix}, or a \code{data.frame}.
Column names are sample names.
Row names are gene names, which can be Ensembl Gene ID, HGNC Symbol,
Entrez Gene ID Ensembl, Transcript ID or Refseq mRNA.
All gene ID will be automatically converted to Ensembl Gene ID.
If you want to retain your gene identifier that is not currently supported,
please specify \code{annotation=NA}. See parameter
\code{annotation} for more information.
Counts of several transcript ID corresponding to same gene will
be added and recorded as the counts of the gene.}

\item{pheno}{Phenotype table corresponding to
the expression matrix.
Row names are sample names, identical to column names of \code{object}.}

\item{colname}{A column name of \code{pheno}.
Cells will be grouped by this column of values.
Thus, differential expression analysis will be performed using this
column of phenotype as independent variables.}

\item{assay}{(Default: NULL) The assay slot to use with your
Bioconductor object.}

\item{data}{A character indicating whether the matrix is
log-transformed or raw counts. By default \code{data="exprs"}.
If the count matrix is raw counts, please specify \code{data="counts"}}

\item{threshold}{(Default: 10) Prefiltering threshold for count row sums.}

\item{add_umap}{Add a UMAP space to the scaffold}

\item{classes}{Cell types to use in plot
For example, set \code{annotation="hgnc_symbol"} will convert the row names
(gene identifiers) of \code{counts_scaffold} to hgnc symbol,
so will the expression data and the resulting PCA scaffold.
If this attempted translation fails, or your desired gene identifier is
not supported (especially when you are analyzing non-human data),
please set \code{annotation="hgnc_symbol"} to avoid translation.
In this case, please manually make sure that the row names
(gene identifiers) of \code{counts_scaffold} and
\code{counts_sample} are the same.}

\item{pval_cutoff}{A cutoff value for p value when selecting
differentially expressed genes. By default \code{pval_cutoff=0.05}.}

\item{lfc_cutoff}{A cutoff value for logFC when selecting
differentially expressed genes. By default \code{lfc_cutoff=2}.}

\item{pca_scale}{A logical variable determining whether to
normalize rows when plotting PCA}

\item{annotation}{Type of gene identifier to use for scaffold.
Currently "ensembl_gene", "ensembl_transcript", "entrez", "hgnc_symbol",
and "refseq_mrna" are supported.}
}
\value{
A scaffold space
}
\description{
This function returns a scaffold space as a list
object that contains all parameters needed for plotting
the PCA/UMAP transformed sample space.
To use prebuilt scaffold space, simply call:
\code{buildScaffold("scaffold_name")}. Get a list of scaffolds with:
\code{\link[spaceRATScaffolds:listScaffolds]{listScaffolds()}}
}
\details{
To build your own scaffold space, pass as arguments a count matrix,
a phenotype table, an a column name of the phenotype table to the function.
Missing values are allowed and automatically processed by the function.
But please make sure that the all missing values are represented by NA,
not "", " "(blank space), "-" or any other character.

When building a user-defined scaffold space,
this function first preprocesses
count matrix by removing genes with total counts less than 10.
Then it performs differential expression analysis to select
differentially expressed genes (DE genes),
subsets the scaffold dataset to contain only the DE genes,
ranks the genes within each sample, finally performs
principle component analysis (PCA) using ranks.
}
\examples{
utils::data("DMAP_exprs", "DMAP_pData", package = "spaceRATScaffolds")
buildScaffold(DMAP_exprs,DMAP_pData,"cell_types", data = "exprs",
pval_cutoff=0.01,pca_scale=TRUE)
}
