% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DMAP_scaffold}
\alias{DMAP_scaffold}
\title{DMAP scaffold}
\format{
a scaffoldSpace
}
\source{
Created from exprs_dmap
}
\usage{
DMAP_scaffold
}
\description{
This scaffold was build from the \code{\link{exprs_dmap}} object.
}
\keyword{datasets}
