% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPlot.R
\name{loadingPlot}
\alias{loadingPlot}
\title{Generate a loading plot of scaffold PCA}
\usage{
loadingPlot(
  scaffold,
  dims = c(1, 2),
  num_genes = 3,
  gene_name = "hgnc_symbol",
  angle = 30,
  df_only = FALSE
)
}
\arguments{
\item{scaffold}{A \code{scaffoldSpace} object returned by
\code{\link{buildScaffold}} function.}

\item{dims}{A numeric vector containing 2 numbers, indicating
which two principle components to plot.}

\item{num_genes}{Optional parameter indicating number of genes to be shown.}

\item{gene_name}{Character indicating which type of gene name to show by
the side of arrows. By default HGNC symbol is shown.}

\item{angle}{A number for the degree of rotation of labels on the plot.}

\item{df_only}{The data.frame for the loading plot will be returned
instead of the plot.}
}
\value{
A data frame indicating the loading scores of the genes
that contribute most to the selected principle components.
The loading plot is printed automatically, thus not returned.
}
\description{
This function takes a previously built scaffold space
and generates a loading plot to visualize the genes
that contribute most to the selected principle components.
}
\examples{
utils::data("DMAP_exprs", "DMAP_pData", package = "spaceRATScaffolds")
scaffold <- buildScaffold(
DMAP_exprs, DMAP_pData, "cell_types", data = "exprs")
loadingPlot(scaffold)
}
