% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPlot.R
\name{loadingPlot}
\alias{loadingPlot}
\title{Generate a loading plot of scaffold PCA}
\usage{
loadingPlot(space, num_genes = 3, gene_name = "hgnc_symbol", angle = 30)
}
\arguments{
\item{space}{A scaffoldSpace object created by function \code{\link{buildScaffold}}.}

\item{num_genes}{An optional parameter indicating number of genes to be shown.}

\item{gene_name}{A character indicating which type of gene name to show by the side of arrows.}

\item{angle}{A number for the degree of rotation of labels on the plot.}
}
\value{
a ggplot object of loading plot, automatically printed
}
\description{
This function takes a scaffoldSpace object and generate a loading plot to visualize the genes that contribute most to selected principle components.
}
\examples{
space <- buildScaffold(exprs_dmap,pData_dmap"cell_types")
loadingPlot(space)
}
