% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadingPlot.R
\name{loadingPlot}
\alias{loadingPlot}
\title{Generate a loading plot of scaffold PCA}
\usage{
loadingPlot(
  space,
  num_genes = 3,
  gene_name = "hgnc_symbol",
  angle = 30,
  df_noly = FALSE
)
}
\arguments{
\item{space}{A scaffoldSpace object created by function \code{\link{buildScaffold}}.}

\item{num_genes}{An optional parameter indicating number of genes to be shown.}

\item{gene_name}{A character indicating which type of gene name to show by the side of arrows. By default HGNC symbol is shown.}

\item{angle}{A number for the degree of rotation of labels on the plot.}
}
\value{
A data frame indicating the loading scores of the genes that contribute most to the selected principle components.
The loading plot is printed automatically, thus not returned.
}
\description{
This function takes a \code{\link{scaffoldSpace}} object and generates a loading plot to visualize the genes that contribute most to the selected principle components.
}
\examples{
space <- buildScaffold(exprs_dmap,pData_dmap"cell_types")
loadingPlot(space)
}
