% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScaffold.R
\name{plotScaffold}
\alias{plotScaffold}
\title{Plot a scaffoldSpace object}
\usage{
plotScaffold(
scaffold,
dimred = "PCA",
title = "Scaffold plot",
plot_mode = "dot",
dims = c(1, 2))
}
\arguments{
\item{scaffold}{A \code{scaffoldSpace} object returned by
\code{\link{buildScaffold}} function.}

\item{dimred}{A character indicating the method for
dimensionality reduction. Currently "PCA" and "UMAP" are supported.
labels to each data point.}

\item{title}{Title for the plot}

\item{plot_mode}{A character indicating whether to add tiny
By default \code{plot_mode="dot"} and tiny labels will not be attached.
If more than 12 cell types are to be displayed, setting
\code{plot_mode="tiny_label"} may yield better visualization.
Shorter names for phenotypes (e.g. cell types) is strongly recommended
in "tiny_label" mode.}

\item{dims}{A numeric vector containing 2 numbers, indicating
which two principle components to plot.}
}
\value{
ggplot object.
}
\description{
This function plot a \code{scaffoldSpace} object created by
\code{\link{buildScaffold}} function.
}
\examples{
scaffold <- buildScaffold("DMAP")
plotScaffold(scaffold, "Scaffold plot title", dimred = "PCA")
}
