% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSample.R
\name{projectSample}
\alias{projectSample}
\title{Project new sample(s) onto existing scaffold PCA plot}
\usage{
projectSample(
  space,
  counts_sample,
  pheno_sample = NULL,
  colname = "cancer_type",
  title = "Samples projected onto scaffold PCA",
  verbose = TRUE,
  annotation = "ensembl_gene"
)
}
\arguments{
\item{space}{A scaffoldSpace object, returned by function \code{\link{buildScaffold}}}

\item{counts_sample}{Expression matrix of new sample, can be either logged or raw data without specification.}

\item{pheno_sample}{Phenotype data corresponding to \code{counts_sample}. If not specified, the output plot will not show legends for new samples.}

\item{colname}{A column name of \code{pheno_sample}. This column of values will be used to annotate projected samples.
This argument should be set together with \code{pheno_sample}.
If \code{pheno_sample} is not specified, this argument will be ignored, and the output plot will not show legends for new samples.}

\item{title}{Title of the plot.}

\item{verbose}{A logical vector indicating whether to report the number of genes imputed to make \code{counts_sample} compatible with \code{counts_scaffold}}

\item{annotation}{Annotation type to use for scaffold. counts_scaffold rownames using alternative identifiers will be translated.
Currently ensembl_gene, entrez, hgnc_symbol, and refseq_mrna are supported. set to "NA", to avoid translation (both scaffold and projected samples must be the same)}
}
\value{
A ggplot object with new samples projected to existing scaffold PCA plot
}
\description{
This function takes in a \code{\link{scaffoldSpace}} objects, subsets the new dataset to contain only the genes that defines the scaffold space,
ranks the subset within new sample(s),
finally projects the new sample(s) onto the existing scaffold PCA plot.
}
\examples{
space <- DMAP_scaffold # or create your own scaffoldSpace
projectSample(space,exprs_ilaria,pData_ilaria,"cancer_type")
}
