% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSample.R
\name{projectSample}
\alias{projectSample}
\title{Project new sample(s) onto existing scaffold PCA plot}
\usage{
projectSample(
    scaffold, sample,
    pheno = NULL,
    colname = NULL,
    assay = NULL,
    dimred = "PCA",
    dims = c(1, 2),
    plot_mode = "dot",
    title = "Samples projected onto scaffold PCA",
    verbose = TRUE,
    annotation = "ensembl_gene"
    )
}
\arguments{
\item{scaffold}{A scaffold space object, returned by function
\code{\link{buildScaffold}}}

\item{sample}{sample data to project. A `matrix`, `data.frame`, or
`SummarizedExperiment`}

\item{pheno}{Phenotype data corresponding to \code{sample}.
If not specified, the output plot will not show legends for new samples.}

\item{colname}{A column name of \code{pheno}.
This column of values will be used to annotate projected samples.
This argument should be set together with \code{pheno}.
If \code{pheno} is not specified, this argument will be ignored,
and the output plot will not show legends for new samples.}

\item{assay}{(Default: NULL) The assay slot to use with your
Bioconductor object.}

\item{dimred}{A character indicating the method for
dimensionality reduction. Currently "PCA" and "UMAP" are supported.
labels to each data point.}

\item{dims}{A numeric vector containing 2 numbers, indicating
which two principle components to plot.}

\item{plot_mode}{A character indicating whether to add tiny
By default \code{plot_mode="dot"} and tiny labels will not be attached.
If more than 12 cell types are to be displayed, setting
\code{plot_mode="tiny_label"} may yield better visualization.
Shorter names for phenotypes (e.g. cell types) is strongly recommended
in "tiny_label" mode.}

\item{title}{Title of the plot.}

\item{verbose}{A logical vector indicating whether to report the number of
genes imputed to make \code{sample} compatible with
\code{scaffold}}

\item{annotation}{Annotation type to use for scaffold.
counts_scaffold rownames using alternative identifiers will be translated.
Currently ensembl_gene, entrez, hgnc_symbol, and refseq_mrna are supported.
set to "NA", to avoid translation
(both scaffold and projected samples must be the same)}
}
\value{
ggplot object with new samples projected to existing scaffold plot
}
\description{
This function takes in a scaffold space objects, subsets
the new dataset to contain only the genes that defines the scaffold space,
ranks the subset within new sample(s),
finally projects the new sample(s) onto the existing scaffold PCA plot.
}
\examples{
utils::data("DMAP_scaffold", "ilaria_counts", "ilaria_pData",
package = "spaceRATScaffolds")
scaffold <- DMAP_scaffold # or create your own scaffoldSpace
projectSample(scaffold,ilaria_counts,ilaria_pData,"cancer_type")
}
