% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projectSample.R
\name{projectSample}
\alias{projectSample}
\title{Project new sample(s) onto the existing scaffold PCA plot}
\usage{
projectSample(
  space,
  exprs_sample,
  pData_sample,
  group_sample,
  title = "Samples projected on scaffold PCA",
  y = "PC2",
  x = "PC1",
  verbose = T
)
}
\arguments{
\item{space}{a scafoldSpace object, returned by function \code{buildScaffold()}}

\item{exprs_sample}{expression matrix of new sample}

\item{pData_sample}{phenotype data corresponding to exprs_sample.}

\item{group_sample}{a column name of pData_sample}

\item{title}{Title of the plot}

\item{verbose}{a logical vector indicating whether to report the number of genes added to eset_sample to make it compatible with eset_scaffold}
}
\value{
a ggplot object with new samples projected to existing scaffold PCA plot
}
\description{
This function takes in a scaffoldSpace objects, subsets the new dataset, ranks the subset, finally projects the new sample(s) onto the existing scaffold PCA plot.
}
\examples{
projectSample(space,exprs_ilaria,pData_ilaria,"cancer_type")
}
