% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_scaffoldSpace.R
\docType{class}
\name{scaffoldSpace-class}
\alias{scaffoldSpace-class}
\alias{scaffoldSpace}
\title{An S4 class to store all the information of scaffold space}
\description{
An S4 class to store all the information of scaffold space
}
\section{Slots}{

\describe{
\item{\code{DEgene}}{a character vector containing names of differentially expressed genes in scaffold dataset}

\item{\code{label}}{a character vector containing labels of each data point}

\item{\code{model}}{a model returned by dimension reduction algorithm. PCA algorithm should return an "prcomp" object. UMAP algorithm should return a "umap" object.}

\item{\code{dims}}{the two indices of reduced dimensions to be plot}

\item{\code{plot_mode}}{a character indicating whether to add tiny label at each data point. Options are "dot" or "tiny_label"}
}}

